Add texture arrays

+++ Mlib/Geometry/Mesh/Uv_Tile.hpp
@@ -11,7 +11,7 @@ struct ManualUvTile {
 struct AutoUvTile {
     FixedArray<float, 2> position;
     FixedArray<float, 2> size;
-    int layer;
+    uint8_t layer;
 };
 
 }
